# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * O Abstrato Doido API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ObraStatus = typeof ObraStatus[keyof typeof ObraStatus];


export const ObraStatus = {
  disponivel: 'disponivel',
  vendido: 'vendido',
} as const;

export interface Obra {
  id: number;
  titulo: string;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  dimensoes?: string | null;
  /** @nullable */
  preco?: string | null;
  status: ObraStatus;
  /** @nullable */
  imagemUrl?: string | null;
  dataCriacao: string;
}

export type ObraInputStatus = typeof ObraInputStatus[keyof typeof ObraInputStatus];


export const ObraInputStatus = {
  disponivel: 'disponivel',
  vendido: 'vendido',
} as const;

export interface ObraInput {
  titulo: string;
  descricao?: string;
  dimensoes?: string;
  preco?: string;
  status: ObraInputStatus;
  imagemUrl?: string;
}

export type ObraUpdateStatus = typeof ObraUpdateStatus[keyof typeof ObraUpdateStatus];


export const ObraUpdateStatus = {
  disponivel: 'disponivel',
  vendido: 'vendido',
} as const;

export interface ObraUpdate {
  titulo?: string;
  descricao?: string;
  dimensoes?: string;
  preco?: string;
  status?: ObraUpdateStatus;
  imagemUrl?: string;
}

export type EventoTipo = typeof EventoTipo[keyof typeof EventoTipo];


export const EventoTipo = {
  futuro: 'futuro',
  passado: 'passado',
} as const;

export interface Evento {
  id: number;
  nome: string;
  /** @nullable */
  local?: string | null;
  dataEvento: string;
  tipo: EventoTipo;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  imagemCapaUrl?: string | null;
  /** @nullable */
  imagensAdicionais?: string | null;
  /** @nullable */
  linkInscricao?: string | null;
}

export type EventoInputTipo = typeof EventoInputTipo[keyof typeof EventoInputTipo];


export const EventoInputTipo = {
  futuro: 'futuro',
  passado: 'passado',
} as const;

export interface EventoInput {
  nome: string;
  local?: string;
  dataEvento: string;
  tipo: EventoInputTipo;
  descricao?: string;
  imagemCapaUrl?: string;
  imagensAdicionais?: string;
  linkInscricao?: string;
}

export type EventoUpdateTipo = typeof EventoUpdateTipo[keyof typeof EventoUpdateTipo];


export const EventoUpdateTipo = {
  futuro: 'futuro',
  passado: 'passado',
} as const;

export interface EventoUpdate {
  nome?: string;
  local?: string;
  dataEvento?: string;
  tipo?: EventoUpdateTipo;
  descricao?: string;
  imagemCapaUrl?: string;
  imagensAdicionais?: string;
  linkInscricao?: string;
}

export interface BlogPost {
  id: number;
  titulo: string;
  conteudo: string;
  /** @nullable */
  imagemCapaUrl?: string | null;
  dataPublicacao: string;
}

export interface BlogPostInput {
  titulo: string;
  conteudo: string;
  imagemCapaUrl?: string;
}

export interface BlogPostUpdate {
  titulo?: string;
  conteudo?: string;
  imagemCapaUrl?: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
}

export type ListObrasParams = {
status?: ListObrasStatus;
limit?: number;
};

export type ListObrasStatus = typeof ListObrasStatus[keyof typeof ListObrasStatus];


export const ListObrasStatus = {
  disponivel: 'disponivel',
  vendido: 'vendido',
} as const;

export type ListEventosParams = {
tipo?: ListEventosTipo;
};

export type ListEventosTipo = typeof ListEventosTipo[keyof typeof ListEventosTipo];


export const ListEventosTipo = {
  futuro: 'futuro',
  passado: 'passado',
} as const;

export type ListBlogParams = {
limit?: number;
};