# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * O Abstrato Doido API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  titulo: string;
  conteudo: string;
  /** @nullable */
  imagemCapaUrl?: string | null;
  dataPublicacao: string;
}