# lib/api-zod/src/generated/types/evento.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * O Abstrato Doido API
 * OpenAPI spec version: 0.1.0
 */
import type { EventoTipo } from './eventoTipo';

export interface Evento {
  id: number;
  nome: string;
  /** @nullable */
  local?: string | null;
  dataEvento: string;
  tipo: EventoTipo;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  imagemCapaUrl?: string | null;
  /** @nullable */
  imagensAdicionais?: string | null;
  /** @nullable */
  linkInscricao?: string | null;
}