# lib/api-zod/src/generated/types/eventoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * O Abstrato Doido API
 * OpenAPI spec version: 0.1.0
 */
import type { EventoInputTipo } from './eventoInputTipo';

export interface EventoInput {
  nome: string;
  local?: string;
  dataEvento: string;
  tipo: EventoInputTipo;
  descricao?: string;
  imagemCapaUrl?: string;
  imagensAdicionais?: string;
  linkInscricao?: string;
}