# lib/api-zod/src/generated/types/eventoUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * O Abstrato Doido API
 * OpenAPI spec version: 0.1.0
 */
import type { EventoUpdateTipo } from './eventoUpdateTipo';

export interface EventoUpdate {
  nome?: string;
  local?: string;
  dataEvento?: string;
  tipo?: EventoUpdateTipo;
  descricao?: string;
  imagemCapaUrl?: string;
  imagensAdicionais?: string;
  linkInscricao?: string;
}